fix: searchboxをコンポーネント単位に分解
