+++ src/components/Header/SearchBox/Pc/index.module.scss
@@ -1,0 +1,6 @@
+.search_box {
+  width: 400px;
+  @media (max-width: 767px) {
+    display: none;
+  }
+}
